Fix SBOM generation for flare in k8s environment
