+++ pkg/collector/corechecks/sbom/processor.go
@@ -10,9 +10,6 @@ package sbom
 import (
 	"context"
 	"errors"
-	"io/fs"
-	"os"
-	"path/filepath"
 	"strings"
 	"time"
 
@@ -21,7 +18,6 @@ import (
 	workloadmeta "github.com/DataDog/datadog-agent/comp/core/workloadmeta/def"
 	"github.com/DataDog/datadog-agent/comp/forwarder/eventplatform"
 	"github.com/DataDog/datadog-agent/pkg/aggregator/sender"
-	"github.com/DataDog/datadog-agent/pkg/config/env"
 
 	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
 	"github.com/DataDog/datadog-agent/pkg/sbom"
@@ -222,39 +218,13 @@ func (p *processor) processHostScanResult(result sbom.ScanResult) {
 	p.queue <- sbom
 }
 
-type relFS struct {
-	root string
-	fs   fs.FS
-}
-
-func newFS(root string) fs.FS {
-	fs := os.DirFS(root)
-	return &relFS{root: "/", fs: fs}
-}
-
-func (f *relFS) Open(name string) (fs.File, error) {
-	if filepath.IsAbs(name) {
-		var err error
-		name, err = filepath.Rel(f.root, name)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return f.fs.Open(name)
-}
-
 func (p *processor) triggerHostScan() {
 	if !p.hostSBOM {
 		return
 	}
 	log.Debugf("Triggering host SBOM refresh")
 
-	scanPath := "/"
-	if hostRoot := os.Getenv("HOST_ROOT"); env.IsContainerized() && hostRoot != "" {
-		scanPath = hostRoot
-	}
-	scanRequest := host.NewScanRequest(scanPath, newFS("/"))
+	scanRequest := host.NewHostScanRequest()
 
 	if err := p.sbomScanner.Scan(scanRequest); err != nil {
 		log.Errorf("Failed to trigger SBOM generation for host: %s", err)
