+++ pkg/sbom/collectors/host/request.go
@@ -7,7 +7,10 @@ package host
 
 import (
 	"io/fs"
+	"os"
+	"path/filepath"
 
+	"github.com/DataDog/datadog-agent/pkg/config/env"
 	"github.com/DataDog/datadog-agent/pkg/sbom/types"
 )
 
@@ -18,9 +21,40 @@ type scanRequest struct {
 	FS   fs.FS
 }
 
+type relFS struct {
+	root string
+	fs   fs.FS
+}
+
+func newFS(root string) fs.FS {
+	fs := os.DirFS(root)
+	return &relFS{root: "/", fs: fs}
+}
+
+func (f *relFS) Open(name string) (fs.File, error) {
+	if filepath.IsAbs(name) {
+		var err error
+		name, err = filepath.Rel(f.root, name)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	return f.fs.Open(name)
+}
+
 // NewScanRequest creates a new scan request
 func NewScanRequest(path string, fs fs.FS) types.ScanRequest {
 	return scanRequest{Path: path, FS: fs}
+}
+
+// NewHostScanRequest creates a new scan request for the root filesystem
+func NewHostScanRequest() types.ScanRequest {
+	scanPath := "/"
+	if hostRoot := os.Getenv("HOST_ROOT"); env.IsContainerized() && hostRoot != "" {
+		scanPath = hostRoot
+	}
+	return NewScanRequest(scanPath, newFS("/"))
 }
 
 // Collector returns the collector name
